Make this work if sizeof(int) < sizeof(intptr_t).
Now we don't truncate nRetVal. Fixes crashes on 64-bit OS X and
other platforms where sizeof(int) < sizeof(intptr_t).

+++ src/MACLib/APEDecompress.cpp
@@ -369,7 +369,7 @@ Get information from the decompressor
 *****************************************************************************************/
 intptr_t CAPEDecompress::GetInfo(APE_DECOMPRESS_FIELDS Field, intptr_t nParam1, intptr_t nParam2)
 {
-    int nRetVal = 0;
+    intptr_t nRetVal = 0;
     BOOL bHandled = TRUE;
 
     switch (Field)
